[ROCM] GPU profiling code refactor - seperarted rocm event tracer and event collector.
